+++ src/components/Dialog/index.js
@@ -1,0 +1,13 @@
+import Fieldset from './Fieldset';
+import Footer from './Footer';
+import Title from './Title';
+import Wrapper from './Wrapper';
+
+const Dialog = {
+  Fieldset,
+  Footer,
+  Title,
+  Wrapper,
+};
+
+export default Dialog;